Move version and testscripts
