+++ cmd/fakermaker/main.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/bluesky-social/indigo/fakedata"
 	"github.com/bluesky-social/indigo/util/cliutil"
-	"github.com/bluesky-social/indigo/version"
+	"github.com/bluesky-social/indigo/util/version"
 
 	_ "github.com/joho/godotenv/autoload"
